refactoring fixes for UnitAxisTags
